+++ packages/datastore-storage-adapter/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@1.1.11...@aws-amplify/datastore-storage-adapter@1.1.12) (2021-11-12)
+
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
+
+## [1.1.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@1.1.10...@aws-amplify/datastore-storage-adapter@1.1.11) (2021-11-09)
+
+### Bug Fixes
+
+- **datastore-storage-adapter:** don't enforce required model fields in sql ([#9166](https://github.com/aws-amplify/amplify-js/issues/9166)) ([fbe1f2e](https://github.com/aws-amplify/amplify-js/commit/fbe1f2ebddb1a155d74e7f45e30a61b95a608af1))
+
 ## [1.1.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@1.1.9...@aws-amplify/datastore-storage-adapter@1.1.10) (2021-10-28)
 
 **Note:** Version bump only for package @aws-amplify/datastore-storage-adapter